LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/script-text-expected.txt


PASS Getter 
PASS Setter (non-empty string) 
FAIL Setter (empty string) assert_equals: expected null but got Text node ""
PASS Setter (null) 
PASS Setter (undefined) 
FAIL Setter (text node reuse) assert_not_equals: got disallowed value Text node "two"
